+++ src/main/java/net/orion/create_limited/Mixins/Implementations/PoweredShaftBlockEntityImpl.java
@@ -17,7 +17,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class PoweredShaftBlockEntityImpl implements DecayingBlockEntity {
 
     @Unique
-    int create_Limited$decay = -1;
+    float create_Limited$decay = -1;
 
     @Unique
     long create_Limited$seconds = 0;
@@ -58,12 +58,12 @@ public abstract class PoweredShaftBlockEntityImpl implements DecayingBlockEntity
     }
 
     @Override
-    public int getCreate_Limited$decay() {
+    public float getCreate_Limited$decay() {
         return this.create_Limited$decay;
     }
 
     @Override
-    public void setCreate_Limited$decay(int i) {
+    public void setCreate_Limited$decay(float i) {
         this.create_Limited$decay = i;
     }
